GUI based on QT/QuDi, opens to main page of QuDi
