+++ config/app_status/status-ManagerGui_gui_man.cfg
@@ -1,0 +1,3 @@
+pos_x: 404
+pos_y: 148
+console_font_size: 10
