+++ config/default.cfg
@@ -7,7 +7,7 @@
 #		  the config line must the form package.module.Class.ClassName
 global:
     # list of modules to load when starting
-    startup: ['man', 'tray', 'tasklogic']
+    startup: ['man', 'tray']
 
     ## For controlling the appearance of the GUI:
     stylesheet: 'qdark.qss'
